+++ Mods ChoGGi/Disable Welcome To Mars Voice/Code/Script.lua
@@ -1,5 +1,8 @@
 -- See LICENSE for terms
 
-function OnMsg.ClassesPostprocess()
-	PopupNotificationPresets.WelcomeGameInfo.voiced_text = false
+-- Workaround for: [mod] Error loading PackedMods/*/Code/Script.lua: PackedMods/*/Code/Script.lua:1: syntax error near '<\218>'
+do
+	function OnMsg.ClassesPostprocess()
+		PopupNotificationPresets.WelcomeGameInfo.voiced_text = false
+	end
 end