+++ Mods ChoGGi/Example Unlock All Buildings/Code/Script.lua
@@ -1,4 +1,7 @@
 -- See LICENSE for terms
 
-OnMsg.CityStart = CheatUnlockAllBuildings
-OnMsg.LoadGame = CheatUnlockAllBuildings
+-- Workaround for: [mod] Error loading PackedMods/*/Code/Script.lua: PackedMods/*/Code/Script.lua:1: syntax error near '<\218>'
+do
+	OnMsg.CityStart = CheatUnlockAllBuildings
+	OnMsg.LoadGame = CheatUnlockAllBuildings
+end