+++ Mods ChoGGi/Older Saves Future Contemporary/Code/Script.lua
@@ -1,5 +1,8 @@
 -- See LICENSE for terms
 
-function OnMsg.LoadGame()
-	g_AccessibleDlc.ariane = true
+-- Workaround for: [mod] Error loading PackedMods/*/Code/Script.lua: PackedMods/*/Code/Script.lua:1: syntax error near '<\218>'
+do
+	function OnMsg.LoadGame()
+		g_AccessibleDlc.ariane = true
+	end
 end